Refactor grid to use flex
